+++ src/storage/ct_classic.cc
@@ -379,14 +379,14 @@ namespace MEDDLY {
           result node(s) are DEAD.
             @param  h   Handle of the entry.
       */
-      // bool entryUnrecoverable(unsigned long h) const;
+      // bool entryUnrecoverable(size_t h) const;
 #endif
 
       /**
           Discard an entry and recycle the memory it used.
             @param  h   Handle of the entry.
       */
-      void discardAndRecycle(unsigned long h);
+      void discardAndRecycle(size_t h);
 
       /**
           Display an entry.
@@ -394,7 +394,7 @@ namespace MEDDLY {
             @param  s         Stream to write to.
             @param  h         Handle of the entry.
       */
-      void showEntry(output &s, unsigned long h) const;
+      void showEntry(output &s, size_t h) const;
 
       /// Display a key.
       void showKey(output &s, const entry_key* k) const;
@@ -946,7 +946,7 @@ void MEDDLY::ct_template<MONOLITHIC, CHAINED>::addEntry(entry_key* key, const en
     }
   }
 
-  unsigned long newsize = tableSize * 2;
+  size_t newsize = tableSize * 2;
   if (newsize > maxSize) newsize = maxSize;
 
   if (CHAINED) {
@@ -1730,7 +1730,7 @@ bool MEDDLY::ct_template<MONOLITHIC, CHAINED>
 
 template <bool MONOLITHIC, bool CHAINED>
 void MEDDLY::ct_template<MONOLITHIC, CHAINED>
-::discardAndRecycle(unsigned long h)
+::discardAndRecycle(size_t h)
 {
   const int SHIFT = (MONOLITHIC ? 1 : 0) + (CHAINED ? 1 : 0);
 
@@ -1870,7 +1870,7 @@ void MEDDLY::ct_template<MONOLITHIC, CHAINED>
 
 template <bool MONOLITHIC, bool CHAINED>
 void MEDDLY::ct_template<MONOLITHIC, CHAINED>
-::showEntry(output &s, unsigned long h) const
+::showEntry(output &s, size_t h) const
 {
 #ifdef INTEGRATED_MEMMAN
   const int* entry = entries+h;
@@ -1933,7 +1933,7 @@ void MEDDLY::ct_template<MONOLITHIC, CHAINED>
                         break;
         case GENERIC:
                         item.G = *((ct_object**)(ptr));
-                        s.put_hex((unsigned long)item.G);
+                        s.put_hex((size_t)item.G);
                         ptr += sizeof(ct_object*) / sizeof(int);
                         break;
         default:
@@ -1974,7 +1974,7 @@ void MEDDLY::ct_template<MONOLITHIC, CHAINED>
                             
         case GENERIC:
                         item.G = *((ct_object**)(ptr));
-                        s.put_hex((unsigned long)item.G);
+                        s.put_hex((size_t)item.G);
                         ptr += sizeof(ct_object*) / sizeof(int);
                         break;
         default:
@@ -2027,7 +2027,7 @@ void MEDDLY::ct_template<MONOLITHIC, CHAINED>
                         s.put(item.D);
                         break;
         case GENERIC:
-                        s.put_hex((unsigned long)item.G);
+                        s.put_hex((size_t)item.G);
                         break;
         default:
                         MEDDLY_DCASSERT(0);
